+++ include/sway/input/keyboard.h
@@ -5,6 +5,27 @@
 
 #define SWAY_KEYBOARD_PRESSED_KEYS_CAP 32
 
+/**
+ * Get modifier mask from modifier name.
+ *
+ * Returns the modifer mask or 0 if the name isn't found.
+ */
+uint32_t get_modifier_mask_by_name(const char *name);
+
+/**
+ * Get modifier name from modifier mask.
+ *
+ * Returns the modifier name or NULL if it isn't found.
+ */
+const char *get_modifier_name_by_mask(uint32_t modifier);
+
+/**
+ * Get an array of modifier names from modifier_masks
+ *
+ * Populates the names array and return the number of names added.
+ */
+int get_modifier_names(const char **names, uint32_t modifier_masks);
+
 struct sway_shortcut_state {
 	/**
 	 * A list of pressed key ids (either keysyms or keycodes),
