+++ sway/commands/bar/modifier.c
@@ -1,8 +1,8 @@
 #include <string.h>
 #include "sway/commands.h"
+#include "sway/input/keyboard.h"
 #include "log.h"
 #include "stringop.h"
-#include "util.h"
 
 struct cmd_results *bar_cmd_modifier(int argc, char **argv) {
 	struct cmd_results *error = NULL;
