+++ sway/commands/floating_modifier.c
@@ -1,7 +1,7 @@
 #include "strings.h"
 #include "sway/commands.h"
 #include "sway/config.h"
-#include "util.h"
+#include "sway/input/keyboard.h"
 
 struct cmd_results *cmd_floating_modifier(int argc, char **argv) {
 	struct cmd_results *error = NULL;
